added device 0x0BDA
I found someone that cracked his open and checked that it's a RTL8811CU on that device ID and from someone ELSE that the vendor ID is 0x0bda. I don't have the device I have no way to test or check the information to see if this is right and messing with drivers this way is not really recommended. If you want to be safe you should confirm everything I said is right before you try it out.